# services/learning_api/openapi.json
{
  "components": {
    "schemas": {
      "AcceptInvitationRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "invitation_code": {
            "maxLength": 300,
            "minLength": 16,
            "title": "Invitation Code",
            "type": "string"
          }
        },
        "required": [
          "invitation_code",
          "display_name"
        ],
        "title": "AcceptInvitationRequest",
        "type": "object"
      },
...
      "CurrentLearnerResponse": {
        "additionalProperties": false,
        "properties": {
          "enrolments": {
            "items": {
              "$ref": "#/components/schemas/EnrolmentResponse"
            },
            "title": "Enrolments",
            "type": "array"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          }
        },
        "required": [
          "profile",
          "enrolments"
        ],
        "title": "CurrentLearnerResponse",
        "type": "object"
      },
...
      "EnrolmentResponse": {
        "additionalProperties": false,
        "properties": {
          "course_key": {
            "title": "Course Key",
            "type": "string"
          },
          "course_revision": {
            "title": "Course Revision",
            "type": "integer"
          },
          "enrolled_at": {
            "format": "date-time",
            "title": "Enrolled At",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "completed",
              "withdrawn"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "course_key",
          "course_revision",
          "status",
          "enrolled_at"
        ],
        "title": "EnrolmentResponse",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[a-z0-9_]+$",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "InvitationAcceptanceResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "const": true,
            "default": true,
            "title": "Accepted",
            "type": "boolean"
          },
          "enrolments": {
            "items": {
              "$ref": "#/components/schemas/EnrolmentResponse"
            },
            "title": "Enrolments",
            "type": "array"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          }
        },
        "required": [
          "profile",
          "enrolments"
        ],
        "title": "InvitationAcceptanceResponse",
        "type": "object"
      },
...
      "ProfileResponse": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "learner_id": {
            "title": "Learner Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "student",
              "content_admin",
              "academic_admin"
            ],
            "title": "Role",
            "type": "string"
          },
          "target_track": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Track"
          }
        },
        "required": [
          "learner_id",
          "email",
          "display_name",
          "role",
          "target_track"
        ],
        "title": "ProfileResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Versioned course and practice API. Lesson bodies remain draft placeholders until reviewed content and video assets are supplied.",
    "title": "ASEAN Academy Learning API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "getCurrentLearner",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentLearnerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get the authenticated learner profile and pinned enrolments",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/onboarding/accept-invitation": {
      "post": {
        "operationId": "acceptBetaInvitation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationAcceptanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept a beta invitation and pin the learner to its course revision",
        "tags": [
          "identity"
        ]
      }
    },
...
  }
}